#1 - Info about disasters added
