+++ public/index.html
@@ -10,6 +10,7 @@
 	<link rel='stylesheet' href='/global.css'>
 	<link rel='stylesheet' href='/bundle.css'>
 	<link href="https://stackpath.bootstrapcdn.com/bootstrap/4.3.1/css/bootstrap.min.css" rel="stylesheet" integrity="sha384-ggOyR0iXCbMQv3Xipma34MD+dH/1fQ784/j6cY/iJTQUOhcWr7x9JvoRxT2MZw1T" crossorigin="anonymous">
+	<link href="https://fonts.googleapis.com/css?family=Asap:400,400i,500,500i,700,700i&amp;subset=latin-ext" rel="stylesheet" />
 	<script defer src='/bundle.js'></script>
 </head>
